refactor: 코드 리뷰 ai 적용 - sse api 로그인 후 사용하도록 필터 제외 - 만료 시간 설정이 잘못 되었던 부분 수정 - event dto 클래스 디렉터리 위치 수정
